+++ src/main/java/Task.java
@@ -1,5 +1,5 @@
 import java.io.Serializable;
-abstract class Task implements Serializable {
+public class Task implements Serializable {
     protected boolean marked;
     protected String description;
     public Task(String description) {
